requirements.txt
asgiref
autopep8==1.5.4
dj-database-url==2.2.0
Django==5.2.6
django-simple-history==3.8.0
django-widget-tweaks==1.4.8
gunicorn==23.0.0
psycopg2-binary==2.9.7
pycodestyle==2.12.0
python-decouple==3.8
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
toml==0.10.2
Unipath==1.1
whitenoise==6.4.0
sentry-sdk==2.8.0
django-select2==8.3.0
djangorestframework==3.16.1
django-filter==24.2
boto3==1.34.137
django-storages[azure]==1.14.4
django-redis==4.12.1
django-ckeditor==6.5.1
pandas==2.2.2
django-pandas==0.6.7
django-anymail[amazon-ses]==12.0
django-environ==0.11.2
setuptools==75.5.0
django-cors-headers==4.6.0
django-guardian==2.4.0
djangorestframework-simplejwt==5.4.0
django-extensions==3.2.3

# Testing
# ------------------------------------------------------------------------------
pytest==8.2.1
pytest-django==4.8.0
pytest-cov==5.0.0
Faker==25.3.0
factory_boy==3.2.1
django-debug-toolbar==4.4.6

# Code quality
# ------------------------------------------------------------------------------
djlint==1.34.1
pre-commit==2.15.0
ruff==0.13.2
django-browser-reload==1.17.0
mkdocs-material==9.5.47
